Title: 세 수의 합과 평균 구하기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231114/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.c
@@ -4,7 +4,7 @@ int main() {
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
     printf("%d",a+b+c);
-    printf("\n%.f",(float)(a+b+c)/3);
+    printf("\n%.f",(a+b+c)/3);
     // 여기에 코드를 작성해주세요.
     return 0;
 }
